Prompt for resetting.

+++ game_tracker/views.py
@@ -29,42 +29,47 @@ def index(request):
     return render(request, 'game_tracker/index.html', context)
 
 def reset(request):
-    Character.objects.all().delete()
+    if "cancel" in request.POST:
+        return redirect('index')
+    elif "reset" in request.POST:
+        Character.objects.all().delete()
 
-    characters = None
-    characters_path = os.path.join(settings.GAME_DIR, 'data', \
-        'characters.json')
-    with open(characters_path) as f:
-        characters_json = f.read()
-        characters = json.loads(characters_json)
+        characters = None
+        characters_path = os.path.join(settings.GAME_DIR, 'data', \
+            'characters.json')
+        with open(characters_path) as f:
+            characters_json = f.read()
+            characters = json.loads(characters_json)
 
-    organized_characters = []
+        organized_characters = []
 
-    current_color = None
-    for character_set in characters:
-        current_color = character_set["color"]
+        current_color = None
+        for character_set in characters:
+            current_color = character_set["color"]
 
-        for character in character_set["characters"]:
-            new_char = Character.objects.create()
+            for character in character_set["characters"]:
+                new_char = Character.objects.create()
 
-            new_char.name = character["name"]
-            new_char.color = current_color
+                new_char.name = character["name"]
+                new_char.color = current_color
 
-            new_char.image = character["name"].split()[0].lower() + ".png"
+                new_char.image = character["name"].split()[0].lower() + ".png"
 
-            new_char.speed_options = character["speed"]
-            new_char.might_options = character["might"]
-            new_char.sanity_options = character["sanity"]
-            new_char.knowledge_options = character["knowledge"]
+                new_char.speed_options = character["speed"]
+                new_char.might_options = character["might"]
+                new_char.sanity_options = character["sanity"]
+                new_char.knowledge_options = character["knowledge"]
 
-            new_char.speed_index = character["default_speed"]
-            new_char.might_index = character["default_might"]
-            new_char.sanity_index = character["default_sanity"]
-            new_char.knowledge_index = character["default_knowledge"]
+                new_char.speed_index = character["default_speed"]
+                new_char.might_index = character["default_might"]
+                new_char.sanity_index = character["default_sanity"]
+                new_char.knowledge_index = character["default_knowledge"]
 
-            new_char.save()
+                new_char.save()
 
-    return redirect('index')
+        return redirect('index')
+    else:
+        return render(request, 'game_tracker/reset.html', None)
 
 def update_attr(request, character_id):
     character = get_object_or_404(Character, pk=character_id)
